finished rewrite of engine core principles
still have bugs to fix. does not compile atm.

+++ REngine3/source/re/app/Application.hpp
@@ -43,7 +43,7 @@ namespace re
 
 	protected:
 		std::string m_appTitle = "REngine3";
-		double m_targetUPS = 60.0;
+		float m_targetUPS = 60.f;
 
 		int m_versionMajor = 1;
 		int m_versionMinor = 0;
